LED_matrix using 74HC595 and 4017
turn on all the LED's and turn on 1 particular LED and Display the
character A

+++ include/delay.h
@@ -1,6 +1,6 @@
 #ifndef __delay_h
 #define __delay_h
 
-void delay_us(int us);	   	// Delay function
+void delay_us(unsigned long int us);	   	// Delay function
 
 #endif
